[IMP] Moved case /api/v2 endpoints into app.blueprints.rest.v2
