+++ source/app/views.py
@@ -97,6 +97,7 @@ from app.blueprints.rest.profile_routes import profile_rest_blueprint
 from app.blueprints.rest.reports_route import reports_rest_blueprint
 from app.blueprints.rest.search_routes import search_rest_blueprint
 from app.blueprints.graphql.graphql_route import graphql_blueprint
+from app.blueprints.rest.v2.case.api_v2_case_routes import api_v2_case_blueprint
 from app.blueprints.rest.v2.case.api_v2_assets_routes import api_v2_assets_blueprint
 from app.blueprints.rest.v2.case.api_v2_ioc_routes import api_v2_ioc_blueprint
 from app.blueprints.rest.v2.case.api_v2_case_tasks_routes import api_v2_tasks_blueprint
@@ -185,6 +186,7 @@ app.register_blueprint(alerts_rest_blueprint)
 app.register_blueprint(rest_api_blueprint)
 app.register_blueprint(demo_blueprint)
 
+app.register_blueprint(api_v2_case_blueprint)
 app.register_blueprint(api_v2_ioc_blueprint)
 app.register_blueprint(api_v2_assets_blueprint)
 app.register_blueprint(api_v2_tasks_blueprint)
